snail-monitoring/requirements.txt
pypylon==4.0.0
pyserial==3.5